restructured repo: elevated nodes subpackage one level up
