+++ examples/complex.py
@@ -1,8 +1,8 @@
 """Reset mechanism demonstration."""
 
-from regelum.environment.node.nodes.base import Node
-from regelum.environment.node.nodes.graph import Graph
-from regelum.environment.node.core.inputs import Inputs
+from regelum.node.nodes.base import Node
+from regelum.node.nodes.graph import Graph
+from regelum.node.core.inputs import Inputs
 from typing import TypeGuard, Optional
