--- regelum/environment/__init__.py
@@ -1,1 +1,0 @@
-"""Environment package."""
