+++ regelum/node/modifiers/zero_order_hold.py
@@ -2,9 +2,9 @@
 
 from typing import Optional, final, Any
 import numpy as np
-from regelum.environment.node.modifiers.base import StepModifier
-from regelum.environment.node.interfaces.node import INode
-from regelum.environment.node.core.variable import Variable
+from regelum.node.modifiers.base import StepModifier
+from regelum.node.interfaces.node import INode
+from regelum.node.core.variable import Variable
 
 
 @final
