+++ regelum/node/nodes/classic_control/controllers/mpc.py
@@ -1,7 +1,7 @@
 """MPC controller demonstration."""
 
-from regelum.environment.node.nodes.base import Node
-from regelum.environment.node.core.variable import Variable
+from regelum.node.nodes.base import Node
+from regelum.node.core.variable import Variable
 from enum import Enum
 import numpy as np
 from typing import Callable, Tuple
