+++ regelum/node/nodes/graph.py
@@ -15,21 +15,21 @@ from copy import deepcopy
 from functools import reduce
 from math import gcd
 
-from regelum.environment.node.nodes.base import Node
-from regelum.environment.node.core.inputs import Inputs, ResolvedInputs
-from regelum.environment.node.interfaces.base import IVariable
-from regelum.environment.node.interfaces.graph import IGraph
-from regelum.environment.node.interfaces.base import IInputs
-from regelum.environment.node.nodes.logging import Clock, StepCounter, Logger
-from regelum.environment.node.nodes.reset import Reset
-from regelum.environment.node.modifiers.reset import ResetOnStep
-from regelum.environment.node.modifiers.zero_order_hold import ZeroOrderHold
-from regelum.environment.node.core.types import ResolveStatus
+from regelum.node.nodes.base import Node
+from regelum.node.core.inputs import Inputs, ResolvedInputs
+from regelum.node.interfaces.base import IVariable
+from regelum.node.interfaces.graph import IGraph
+from regelum.node.interfaces.base import IInputs
+from regelum.node.nodes.logging import Clock, StepCounter, Logger
+from regelum.node.nodes.reset import Reset
+from regelum.node.modifiers.reset import ResetOnStep
+from regelum.node.modifiers.zero_order_hold import ZeroOrderHold
+from regelum.node.core.types import ResolveStatus
 from regelum.utils.logger import logger
 from regelum.utils import find_scc
 
 if TYPE_CHECKING:
-    from regelum.environment.node.nodes.parallel import ParallelGraph
+    from regelum.node.nodes.parallel import ParallelGraph
 
 
 class Graph(Node, IGraph[Node]):
@@ -794,7 +794,7 @@ class Graph(Node, IGraph[Node]):
 
     def parallelize(self, **kwargs: Any) -> ParallelGraph:
         """Convert to parallel execution mode."""
-        from regelum.environment.node.nodes.parallel import ParallelGraph
+        from regelum.node.nodes.parallel import ParallelGraph
 
         return ParallelGraph(self.nodes, self.debug, **kwargs)
