+++ tests/test_pendulum_integration.py
@@ -1,8 +1,8 @@
 import pytest
 import numpy as np
 from typing import cast
-from regelum.environment.node.nodes.base import Node
-from regelum.environment.node.nodes.graph import Graph
+from regelum.node.nodes.base import Node
+from regelum.node.nodes.graph import Graph
 from regelum.utils import rg
